requirements.txt
faiss-gpu==1.7.2
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
stereopy==0.9.0
tqdm==4.60.0
rpy2==3.5.7
scanpy==1.9.1
jinja2==3.0.3
